API reference

Public API

Scheme parameters

SGFHE.Params — Type

FHE scheme parameters.

Fields

n :: Int - polynomial length.

Params(n::Int; rlwe_type=nothing, mod_repr=nothing)

n: polynomial length. Must be a power of 2, >=64.

rlwe_type: an unsigned integer type used for bootstrapping (UInt or MLUInt of a large enough size).

mod_repr: the modulo integer representation, must be one of DarkIntegers.ModUInt, DarkIntegers.MgModUInt

Key generation

SGFHE.PrivateKey — Type

PrivateKey(params::Params, rng::AbstractRNG)

Creates the FHE private key for the FHE parameters.

SGFHE.PublicKey — Type

PublicKey(rng::AbstractRNG, sk::PrivateKey)

Creates the FHE public key based on the given private key.

SGFHE.BootstrapKey — Type

BootstrapKey(rng::AbstractRNG, sk::PrivateKey)

Creates the FHE bootstrap key based on the given private key.

Encryption

SGFHE.encrypt — Function

encrypt(key::PrivateKey, rng::AbstractRNG, message::AbstractArray{Bool, 1})

Encrypts a message with the private key. The message must have length equal to the polynomial length n (see Params.n). Returns a PackedCiphertext object.

encrypt(key::PublicKey, rng::AbstractRNG, message::AbstractArray{Bool, 1})

Encrypts a message with the public key. The message must have length equal to the polynomial length n (see Params.n). Returns a PackedCiphertext object.

SGFHE.encrypt_optimal — Function

encrypt_optimal(key::PrivateKey, rng::AbstractRNG, message::AbstractArray{Bool, 1})

Encrypts a message with the private key producing a space-optimal representation (6 bits per bit of the message). The message must have length equal to the polynomial length n (see Params.n). Returns a PrivateEncryptedCiphertext object.

encrypt_optimal(key::PublicKey, rng::AbstractRNG, message::AbstractArray{Bool, 1})

Encrypts a message with the public key producing a space-optimal representation (6 bits per bit of the message). The message must have length equal to the polynomial length n (see Params.n). Returns a PublicEncryptedCiphertext object.

Decryption

SGFHE.decrypt — Function

decrypt(key::PrivateKey, ct::Union{Ciphertext, PackedCiphertext})

Decrypts an n-bit message (see Params.n) from an RLWE using the private key. Returns an Array{Bool, 1} object.

decrypt(key::PrivateKey, ct::EncryptedBit)

Decrypts an single-bit message from an LWE using the private key. Returns a Bool object.

Ciphertext transformations

SGFHE.normalize_ciphertext — Function

normalize_ciphertext(ct::PrivateEncryptedCiphertext)

Converts a space-optimal private-encrypted ciphertext into a generic PackedCiphertext object.

normalize_ciphertext(ct::PublicEncryptedCiphertext)

Converts a space-optimal public-encrypted ciphertext into a generic PackedCiphertext object.

SGFHE.split_ciphertext — Function

split_ciphertext(ct::Union{Ciphertext, PackedCiphertext})

Splits an RLWE ciphertext (encrypting n bits) into n separate EncryptedBit objects. Returns an Array{EncryptedBit, 1}.

SGFHE.pack_encrypted_bits — Function

pack_encrypted_bits(
    bkey::BootstrapKey, rng::Union{AbstractRNG, Nothing},
    enc_bits::AbstractArray{EncryptedBit, 1})

Converts an array of n EncryptedBit objects (see Params.n) into an RLWE Ciphertext. If nothing is given for rng, the result will be deterministic.

Bootstrap

SGFHE.bootstrap — Function

bootstrap(
    bkey::BootstrapKey, rng::Union{AbstractRNG, Nothing},
    enc_bit1::EncryptedBit, enc_bit2::EncryptedBit)

Based on EncryptedBit objects encrypting bits y1 and y2, produces a tuple of three EncryptedBit objects encrypting y1 & y2, y1 | y2 and xor(y1, y2). If nothing is given for rng, the result will be deterministic.

Internals

Returned types

SGFHE.PrivateEncryptedCiphertext — Type

A space-optimal representation of n bits encrypted with a private key (taking 6n bits in total).

SGFHE.PublicEncryptedCiphertext — Type

A space-optimal representation of n bits encrypted with a public key (taking (10 + log2(n))n bits in total).

SGFHE.PackedCiphertext — Type

A packed RLWE ciphertext encrypting an n-bit message that can be produced by an initial encryption with a private or a public key. Takes 2 * t * n bits, where t is the bit size of the integer type used.

SGFHE.Ciphertext — Type

An RLWE ciphertext encrypting an n-bit message that can be produced by joining together n EncryptedBit objects. Takes 16 * t * n bits, where t is the bit size of the integer type used.

SGFHE.EncryptedBit — Type

An LWE ciphertext encrypting a single bit.

Internal functions

SGFHE.find_modulus — Function

Find a residue ring modulus q that:

qmin <= q <= qmax

q is prime

q - 1 is a multiple of n

SGFHE.prng_expand — Function

prng_expand(seq::BitArray{1}, factor::Int)

Expands a sequence of n bits into a pseudo-random array of n * factor bits deterministically, representing each factor bits as an integer.

SGFHE.reduce_modulus — Function

function reduce_modulus(
    new_repr, new_base_type, new_modulus::Unsigned,
    x::Union{<:AbstractModUInt, Polynomial{<:AbstractModUInt}},
    floor_result::Bool=false,
    new_max::Union{Nothing, Unsigned}=nothing)

Reduces the modulus of the given value or all coefficients of the given polynomial, simultaneously casting them to the residue ring representation new_repr and type new_base_type. If floor_result is true, use flooring when rescaling, otherwise use rounding. If new_max is nothing, it is taken equal to new_modulus.

SGFHE.flatten — Function

$\triangleleft$ operator in the paper.

Returns an l-tuple b such that sum(b .* B.^(0:l-1)) == a. (with all the operations performed in the residue ring a belongs to).

Assumes B^l <= q where q is the modulus of a.

flatten(rng::Nothing, a::AbstractModUInt, ::Val{B}, ::Val{l})

The result is deterministic, and each element of the returned tuple -B/2 < b[i] <= B/2 (where the comparisons are modulo modulus(a)).

flatten(rng::AbstractRNG, a::AbstractModUInt, ::Val{B}, ::Val{l})

The result is randomized, and each element of the returned tuple -2B < b[i] <= 2B (where the comparisons are modulo modulus(a)).

SGFHE.flatten_poly — Function

function flatten_poly(
    rng::Union{AbstractRNG, Nothing},
    a::Polynomial{<:AbstractModUInt}, B_val::Val{B}, l_val::Val{l})

Applies flatten to each of the coefficients of the polynomial, and return a tuple of l polynomials, where the i-th polynomial is created out of i-th elements of the returned tuples.

SGFHE.external_product — Function

external_product(
    rng::Union{AbstractRNG, Nothing},
    a::Polynomial{T, N}, b::Polynomial{T, N},
    A::Array{Polynomial{T, N}, 2}, B_val::Val, l_val::Val)

$\odot$ operator in the paper. Returns [flatten(a); flatten(b)] * A. If rng is nothing, deterministic flattening is used.

SGFHE.shortened_external_product — Function

shortened_external_product(
    rng::Union{AbstractRNG, Nothing},
    a::Polynomial{T, N}, A::Array{Polynomial{T, N}, 2}, B_val::Val, l_val::Val)

$\odot$ operator in the paper. Returns flatten(a) * A[l+1:2l,:]. If rng is nothing, deterministic flattening is used.

SGFHE.extract — Function

extract(a::Polynomial, i::Integer, n::Integer)

Extract a sub-array of length n from the polynomial coefficients.